// attendance-and-schedule/src/types/notification.ts
import { notifications } from "@/database/schemas/notification";

export type Notification = typeof notifications.$inferSelect;
export type NotificationRecipient = {
  id: number;
  isRead: boolean | null;
  notification: {
    title: string;
    message: string;
    priority: string | null;
  };
};
export type MarkAsRead = {
  id: number;
  isRead: boolean | null;
};